feat: integrate think_tool and J-Quants API tools for enhanced research capabilities
