+++ src/open_deep_research/utils.py
@@ -18,49 +18,11 @@ from mcp import McpError
 from langchain_mcp_adapters.client import MultiServerMCPClient
 from open_deep_research.state import Summary, ResearchComplete
 from open_deep_research.configuration import SearchAPI, Configuration
+from open_deep_research.tools import think_tool, get_recent_stock_price_tool, get_financial_statements_tool
 from open_deep_research.prompts_jp import summarize_webpage_prompt
 
 
 ##########################
-# Reflection Tool Utils
-##########################
-
-@tool(description="Strategic reflection tool for research planning and evaluation - MUST USE before and after each research step")
-def think_tool(reflection: str) -> str:
-    """Tool for strategic reflection on research progress and decision-making.
-    
-    🔥 CRITICAL: This tool MUST be used at key decision points in research workflow:
-    1. BEFORE starting research: Plan strategy and approach
-    2. AFTER each tool execution: Evaluate results and plan next steps
-    3. BEFORE concluding: Assess if information is sufficient for complete answer
-
-    When to use (MANDATORY):
-    - Before any research: What is my strategic approach and priority areas?
-    - After receiving search results: What key information did I find? What gaps remain?
-    - Before deciding next steps: Do I have enough to answer comprehensively?
-    - When assessing research gaps: What specific information am I still missing?
-    - Before concluding research: Can I provide a complete answer now?
-
-    Reflection should address:
-    1. Strategic planning - What is my approach and priority for this research step?
-    2. Analysis of current findings - What concrete information have I gathered?
-    3. Gap assessment - What crucial information is still missing?
-    4. Quality evaluation - Do I have sufficient evidence/examples for a good answer?
-    5. Strategic decision - Should I continue searching or provide my answer?
-
-    Args:
-        reflection: Your detailed reflection on research strategy, progress, findings, gaps, and next steps
-
-    Returns:
-        Confirmation that reflection was recorded for decision-making
-    """
-    # ログを出力
-    print(f"\n🧠 THINK_TOOL CALLED:")
-    print(f"📝 Reflection: {reflection}")
-    
-    return f"Reflection recorded: {reflection}"
-
-
 ##########################
 # Tavily Search Tool Utils
 ##########################
@@ -312,154 +274,6 @@ async def load_mcp_tools(
         tools.append(wrap_mcp_authenticate_tool(tool))
     return tools
 
-
-# ##########################
-# # Custom Tools
-# ##########################
-from open_deep_research.jquants_api import JQuantsAPI
-
-JQUANTS_FINANCIAL_DESCRIPTION = (
-    "J-Quants APIを使って企業コードと年度から財務情報（売上高、営業利益、当期純利益など）を取得します。過去5年分のデータが取得可能です。"
-)
-
-# Rate limiting management
-_last_api_call_time = 0
-_min_delay_between_calls = 2.0  # 最小2秒間隔
-
-async def rate_limit_delay():
-    """API呼び出し間に適切な遅延を挿入"""
-    global _last_api_call_time
-    current_time = time.time()
-    time_since_last_call = current_time - _last_api_call_time
-    
-    if time_since_last_call < _min_delay_between_calls:
-        delay = _min_delay_between_calls - time_since_last_call
-        # ランダムな要素を追加してリクエストを分散
-        delay += random.uniform(0.5, 1.5)
-        await asyncio.sleep(delay)
-    
-    _last_api_call_time = time.time()
-
-def remove_empty_values(data):
-    """
-    辞書から値が空（None, 空文字, 空リスト, 空辞書）のキーを再帰的に削除する
-    
-    Args:
-        data: 処理対象のデータ（辞書、リスト、その他）
-    
-    Returns:
-        空値が削除されたデータ
-    """
-    if isinstance(data, dict):
-        cleaned = {}
-        for key, value in data.items():
-            # 再帰的に処理
-            cleaned_value = remove_empty_values(value)
-            # 空でない値のみ保持
-            if cleaned_value not in (None, "", [], {}):
-                cleaned[key] = cleaned_value
-        return cleaned
-    elif isinstance(data, list):
-        # リストの場合、各要素を再帰的に処理
-        cleaned_list = []
-        for item in data:
-            cleaned_item = remove_empty_values(item)
-            if cleaned_item not in (None, "", [], {}):
-                cleaned_list.append(cleaned_item)
-        return cleaned_list
-    else:
-        # プリミティブ型はそのまま返す
-        return data
-
-@tool(description=JQUANTS_FINANCIAL_DESCRIPTION)
-async def get_financial_statements_tool(
-    code: str,
-    year: Optional[int] = None,
-    config: RunnableConfig = None
-) -> Dict[str, Any]:
-    """
-    J-Quants APIを使って財務情報を取得します。
-
-    Args:
-        code (str): 企業コード
-        year (Optional[int]): 年度（指定しない場合は最新）
-    Returns:
-        財務情報の辞書
-    """
-    # レート制限対応の遅延
-    await rate_limit_delay()
-    
-    api = JQuantsAPI()
-    raw_data = api.get_financial_statements(code, year)
-    # 空の値を削除
-    cleaned_data = remove_empty_values(raw_data)
-    return cleaned_data
-
-JQUANTS_STOCK_PRICE_DESCRIPTION = (
-    "【最優先ツール】J-Quants APIを使って企業コードから現在の株価情報を取得します。"
-    "株式分析では必ず最初に実行してください。直近1週間の株価データが取得できます。"
-)
-
-@tool(description=JQUANTS_STOCK_PRICE_DESCRIPTION)
-async def get_recent_stock_price_tool(
-    code: str,
-    config: RunnableConfig = None
-) -> Dict[str, Any]:
-    """
-    【重要】株式分析で最初に実行すべきツール
-    J-Quants APIを使って企業コードから現在の株価情報を取得します。
-    投資判断には現在の株価情報が必須です。
-
-    Args:
-        code (str): 企業コード（4桁の数字）例：7203（トヨタ）、8697（楽天）
-    Returns:
-        現在の株価情報を含む辞書（終値、出来高、高値・安値等）
-    """
-    from datetime import datetime, timedelta
-    
-    # 企業コードの検証
-    if not code.isdigit() or len(code) != 4:
-        return {
-            "error": f"無効な企業コード: {code}（4桁の数字である必要があります）",
-            "valid_format": "例: 7203（トヨタ）, 6502（東芝）, 9984（ソフトバンク）"
-        }
-    
-    try:
-        # レート制限対応の遅延
-        await rate_limit_delay()
-        
-        # 直近1週間の日付を計算（土日を考慮して営業日のみ）
-        end_date = datetime.now()
-        start_date = end_date - timedelta(days=10)  # 土日を考慮して10日前から
-        
-        date_from = start_date.strftime("%Y-%m-%d")
-        date_to = end_date.strftime("%Y-%m-%d")
-        
-        api = JQuantsAPI()
-        result = api.get_stock_price(code=code, date_from=date_from, date_to=date_to)
-        
-        # 結果に企業コード情報を追加
-        result["requested_code"] = code
-        result["date_range"] = f"{date_from} to {date_to}"
-        
-        return result
-        
-    except Exception as e:
-        return {
-            "error": f"株価取得エラー: {str(e)}",
-            "code": code,
-            "suggestion": "企業コードが正しいか、または企業が東証上場しているかを確認してください"
-        }
-# @tool(description="Query a database for research data")
-# async def database_query(
-#     query: str,
-#     table: str = "research_data",
-#     config: RunnableConfig = None
-# ) -> str:
-#     """Execute a database query to retrieve research data."""
-#     # Your database connection and query logic here
-#     # This is a placeholder implementation
-#     return f"Database query result for: {query} from table {table}"
 
 ##########################
 # Tool Utils
